speech_modulation/requirements.txt
# AssemblyAI for speech analysis
assemblyai>=0.48.4

# Environment variable management
python-dotenv==1.0.0



